refactor: extract search.js from app.js and rename it to main.js
