+++ static/index.html
@@ -172,6 +172,6 @@
     </div>
   </div>
 
-  <script type="module" src="/static/app.js"></script>
+  <script type="module" src="/static/main.js"></script>
 </body>
 </html>
